tests refactoring and added storage_proof tests
